Moved CIGAR processing function to common.c
